requirements.txt
fastapi>=0.115.12
python-multipart>=0.0.20
uvicorn>=0.34.2
pydantic-settings>=2.9.1
python-dotenv>=1.1.0
syft-core>=0.2.3
git+https://github.com/OpenMined/rds.git@rasswanth/add-enclave-enc
